task(Profile): Tab content ordered in blocks
