+++ interfaces/index.ts
@@ -116,3 +116,23 @@ export type CustomResponse<DataType> = {
   hasPreviousPage: boolean;
   data: DataType[]
 }
+
+export const NFT_OWNED_TAB = 'My NFTs';
+export const NFT_ON_SALE_TAB = 'On sale';
+export const NFT_NOT_FOR_SALE_TAB = 'Not for sale';
+export const NFT_CREATED_TAB = 'Created';
+export const NFT_LIKED_TAB = 'Liked';
+export const FOLLOWERS_TAB = 'Followers';
+export const FOLLOWED_TAB = 'Following';
+
+const ALL_TABS_ID = [
+  NFT_OWNED_TAB,
+  NFT_ON_SALE_TAB,
+  NFT_NOT_FOR_SALE_TAB,
+  NFT_CREATED_TAB,
+  NFT_LIKED_TAB,
+  FOLLOWERS_TAB,
+  FOLLOWED_TAB,
+] as const;
+
+export type TabsIdType = typeof ALL_TABS_ID[number];
